Make `MongoArray` internal
