+++ src/main/java/com/mongodb/hibernate/internal/jdbc/ArrayAdapter.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.mongodb.hibernate.jdbc;
+package com.mongodb.hibernate.internal.jdbc;
 
 import java.sql.Array;
 import java.sql.ResultSet;
